Fixes a broken shake bug
When the user needs to sake (event added) and then closes the story
(shake listener removed) then it won't ever allow it to work.

Now you can shake anytime an if it need to unlock something then it does
otherwise it stays around waiting for when a shake does have.